% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_unidad3.R
\name{elipse}
\alias{elipse}
\title{Punto dentro de la elipse}
\usage{
elipse(x = 0, y = 0)
}
\arguments{
\item{x, y}{coordenadas, números reales. Por defecto valen 0.}
}
\value{
valor lógico indicando si el punto está dentro de la elipse descripta.
}
\description{
Determina si un punto está contenido dentro de la elipse definida por la fórmula
(x - 6)^2 / 36 + (y + 4)^2 / 16 = 1
}
\details{
Por defecto se evalúa al origen de coordenadas.
}
\examples{
elipse(3, 7)
elipse(6, -4)
elipse()

}
