% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_unidad3.R
\name{suma_piramide}
\alias{suma_piramide}
\title{Suma de fila de la pirámide}
\usage{
suma_piramide(n)
}
\arguments{
\item{n}{Número natural}
}
\value{
suma de los números ubicados en la n-ésima fila de la pirámide.
}
\description{
La pirámide se arma con números impares, empezando con el 1 en la cima, el 3 y 5
en la segunda fila, y así sucesivamente. Devuelve la suma de los números ubicados
en la fila ingresada.
}
\details{
Se puede demostrar que la suma de los números en la n-ésima fila es igual a
n al cubo.
}
\examples{
suma_piramide(1)
suma_piramide(2)
suma_piramide(3)

}
