% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_unidad3.R
\name{suma_secuencia}
\alias{suma_secuencia}
\title{Suma de una secuencia de números enteros}
\usage{
suma_secuencia(a, b)
}
\arguments{
\item{a, b}{números enteros}
}
\value{
suma de la secuencia
}
\description{
Calcula la suma de un rango de números enteros, incluyendo los extremos.
}
\examples{
suma_secuencia(1, 3)
suma_secuencia(-2, 3)
suma_secuencia(-3, -5)
suma_secuencia(3, 3)

}
