% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_unidad3.R
\name{triangulos}
\alias{triangulos}
\title{Clasificación de un triángulo}
\usage{
triangulos(a, b, c)
}
\arguments{
\item{a, b, c}{números reales positivos}
}
\value{
valor carácter que indica el tipo de triángulo.
}
\description{
Clasifica a un triángulo según las longitudes de sus lados, en escaleno, isósceles
o equilátero.
}
\details{
Se evalúa la desigualdad triangular. Si las medidas de los lados no corresponden
a un triángulo, la función devuelve "no es un triángulo".
}
\examples{
triangulos(2, 3, 4)
triangulos(2, 3, 10)

}
